Add getData function to get all elements from storage on load
